backend/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
boto3==1.40.6
botocore==1.40.6
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
databricks-sdk==0.62.0
docker==7.1.0
fastapi==0.116.1
Flask==3.1.1
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.5
minio==7.2.16
mlflow==3.2.0
mlflow-skinny==3.2.0
mlflow-tracing==3.2.0
numpy==2.3.2
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.16.1
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlparse==0.5.3
starlette==0.47.2
threadpoolctl==3.6.0
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
waitress==3.0.2
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
zipp==3.23.0
